@@ -1,12 +1,14 @@
-import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
-import store from "./store";
+import './assets/css/tailwind.css';
 
-Vue.config.productionTip = false;
+import { createApp } from 'vue';
 
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount("#app");
+import App from './App.vue';
+import { router } from './router';
+import { store } from './store';
+
+const app = createApp(App);
+
+app.use(router);
+app.use(store);
+
+app.mount('#app');
