@@ -1,28 +1,27 @@
-import Vue from "vue";
-import VueRouter from "vue-router";
-import Home from "../views/Home.vue";
+import { createRouter, createWebHistory } from 'vue-router';
 
-Vue.use(VueRouter);
+import Home from '../views/Home.vue';
 
-const routes = [
+export const routerHistory = createWebHistory();
+
+export const routes = [
   {
-    path: "/",
-    name: "Home",
+    path: '/',
+    name: 'Home',
     component: Home
   },
   {
-    path: "/about",
-    name: "About",
+    path: '/about',
+    name: 'About',
     // route level code-splitting
     // this generates a separate chunk (about.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
     component: () =>
-      import(/* webpackChunkName: "about" */ "../views/About.vue")
+      import(/* webpackChunkName: "about" */ '../views/About.vue')
   }
 ];
 
-const router = new VueRouter({
+export const router = createRouter({
+  history: routerHistory,
   routes
 });
-
-export default router;
